Refactor OverlayEntry to separate widget
